% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age_five}
\alias{age_five}
\title{Create 5-year age labels}
\usage{
age_five(min = 0, max = 100, open = NULL)
}
\arguments{
\item{min}{Minimum age. Defaults to 0.}

\item{max}{Maximum age for closed age groups.
Defaults to 100.}

\item{open}{Whether the last age group is
"open", ie has no upper limit.}
}
\value{
A character vector of length
\code{(max - min)/5 + open}.
}
\description{
Create labels for 5-year age groups.
}
\details{
\code{min} and \code{max} refer to
exact ages. When \code{open} is \code{TRUE},
the age groups are defined as
\code{[min, min+5), [min+5, min+10), \dots,
[max-5, max), [max, Inf)}.
When \code{open} is \code{FALSE},
the age groups are defined as
\code{[min, min+5), [min+5, min+10), \dots,
[max-5, max)}.

\code{open} defaults to \code{TRUE} when
\code{min} equals zero, and to \code{FALSE}
otherwise.
}
\examples{
age_five()
age_five(max = 25)
age_five(max = 25, open = FALSE)
age_five(min = 15, max = 40)
}
\seealso{
\code{\link{age_single}} for 1-year
age groups and \code{\link{age_lt}} for
life table age groups.
}
