% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age_groups}
\alias{age_groups}
\title{Infer type of age labels}
\usage{
age_groups(x)
}
\arguments{
\item{x}{A vector of age labels}
}
\value{
\code{"single"}, \code{"five"}, or \code{"lt"}.
}
\description{
Determine whether a set of age labels
are one-year, five-year, or
life table age groups.
}
\details{
The valid types of age labels are:
\itemize{
\item \code{"single"}. One-year age groups, eg
\code{"0"} or \code{"55"}, and possibly
an open age group, eg \code{"90+"}.
\item \code{"five"}. Five-year age groups, eg
\code{"0-4"} or \code{"55-59"}, and possibly
an open age group, eg \code{"100+"}.
\item \code{"lt"}. Life table age groups, eg
\code{"0"}, \code{"1-4"}, \code{"5-9"},
\code{"55-59"}, or \code{"80+"}.
}

If \code{x} does not fit any of these
descriptions, \verb{then age_groups()} throws
an error.

If \code{x} could belong to
more than one type, then \code{age_groups()}
prefers \code{"single"} to \code{"five"} and \code{"lt"},
and prefers \code{"five"} to \code{"lt"}.
}
\examples{
age_groups(c("5-9", "0-4", "100+"))
age_groups(c("2", "5", "1"))
age_groups(c("0", "1-4"))

## could be any "single" or "lt"
age_groups("0")

## could be "five" or "lt"
age_groups("80-84")
}
