% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age_lt}
\alias{age_lt}
\title{Create life table age labels}
\usage{
age_lt(max = 100)
}
\arguments{
\item{max}{Maximum age for closed age groups.
Defaults to 100.}
}
\value{
A character vector.
}
\description{
Create labels for 'abridged' life table
age groups, ie \code{"0", "1-4", "5-9",
"10-14", \dots, "<max>+"}.
}
\examples{
age_lt()
age_lt(max = 25)
}
\seealso{
\code{\link{age_single}} for 1-year
age groups and \code{\link{age_five}} for
5-year age groups.
}
