% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age_single}
\alias{age_single}
\title{Create one-year age labels}
\usage{
age_single(min = 0, max = 100, open = NULL)
}
\arguments{
\item{min}{Minimum age. Defaults to 0.}

\item{max}{Maximum age for closed age groups.
Defaults to 100.}

\item{open}{Whether the last age group is
"open", ie has no upper limit.}
}
\value{
A character vector of length
\code{max - min + open}.
}
\description{
Create labels for one-year age groups.
}
\details{
\code{min} and \code{max} refer to
exact ages. When \code{open} is \code{TRUE},
the age groups are defined as
\code{[min, min+1), [min+1, min+2), \dots,
[max-1, max), [max, Inf)}.
When \code{open} is \code{FALSE},
the age groups are defined as
\code{[min, min+1), [min+1, min+2), \dots,
[max-1, max)}.

\code{open} defaults to \code{TRUE} when
\code{min} equals zero, and to \code{FALSE}
otherwise.
}
\examples{
age_single()
age_single(max = 5)
age_single(max = 5, open = FALSE)
age_single(min = 15, max = 40)
}
\seealso{
\code{\link{age_five}} for 5-year
age groups and \code{\link{age_lt}} for
life table age groups.
}
