% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{booth_standard}
\alias{booth_standard}
\title{Booth Standard}
\format{
A tibble with 8 rows and the following columns:
\itemize{
\item \code{age} Five-year age group, from \code{"10-14"} to \code{"45-49"}
\item \code{value} Proportion of total fertility occurring within
the age group
}
}
\source{
Booth H (1984) Transforming Gompertz's function
for fertility analysis: The development of a standard
for the relational Gompertz function.
\emph{Population Studies} 38(3): 495-506.
}
\usage{
booth_standard
}
\description{
A "standard" distribution for age-specific fertility
rates proposed in Booth (1984), and widely used for
estimating fertility rates in settings with high
fertility and limited data.
}
\details{
The key input for \code{booth_standard}
is 'Yx' from table 2 of Booth (1984).
}
\keyword{datasets}
