% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-functions.R
\name{check_equal_length}
\alias{check_equal_length}
\title{Check that Arguments have Same Length}
\usage{
check_equal_length(x, y, nm_x, nm_y)
}
\arguments{
\item{x, y}{Arguments to compare}

\item{nm_x, nm_y}{Names to use in error message}
}
\value{
'TRUE', invisibly.
}
\description{
Check that \code{x} and \code{y} have the same length.
}
\examples{
x <- 1:3
y <- 3:1
check_equal_length(x = x,
                   y = y,
                   nm_x = "x",
                   nm_y = "y")
}
