% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetab.R
\name{check_no_overlap_colnums}
\alias{check_no_overlap_colnums}
\title{Check that column numbers do not overlap}
\usage{
check_no_overlap_colnums(x)
}
\arguments{
\item{A}{named list of integer vectors.}
}
\value{
\code{TRUE}, invisibly
}
\description{
Check that column numbers do not overlap
}
