% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{clean_age}
\alias{clean_age}
\title{Clean age group labels}
\usage{
clean_age(x, factor = TRUE)
}
\arguments{
\item{x}{A vector.}

\item{factor}{Whether the return value
should be a factor.}
}
\value{
If \code{factor} is \code{TRUE},
then \code{clean_age} returns a factor;
otherwise it returns a character vector.
}
\description{
Convert age group labels to a standard format.
\describe{
  \item{Step 1} Tidy and translate text,
    eg convert \code{"20 to 24 years"} to
    \code{"20-24"}, convert \code{"infant"} to
    \code{"0"}, and convert \code{"100 or more"} to
    \code{"100+"}.
  \item{Step 2} Check whether the resulting
    labels could have been produced by
    \code{\link{age_labels}}. If not, throw an error.
  \item{Step 3} If \code{factor} is \code{TRUE}
    (the default), then return a factor where the levels
    include all intermediate age groups. Otherwise
    return a character vector.
}
}
\details{
\code{clean_age} allows for two special cases:
  \describe{
    \item{5-year} Labels consisting entirely of
      multiples of 5, with a maximum of at least
      50, are assumed to denote 5-year age groups.
   \item{Life table} Labels consisting entirely
      of 0, 1, and multiples of 5, with a maximum
      of at least 50, are assumed to denote
      life table age groups.
}
}
\examples{
## five-year, factor
clean_age(c("65 and over",
            "10--14",
            "20-24 years"))

## life table
clean_age(c("0",
            "30-34",
            "10--14",
            "1-4 years"))

## single
clean_age(c("60", "90plus"))

## non-factor
clean_age(c("60", "90plus"),
          factor = FALSE)
}
\seealso{
\code{\link{age_labels}}
}
