% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{collapse_age}
\alias{collapse_age}
\title{Collapse age groups}
\usage{
collapse_age(x, type_to = c("five", "lt"))
}
\arguments{
\item{x}{A vector of age labels}

\item{type_to}{Type of age classification
to convert to: \code{"five"} or \code{"lt"}.
Defaults to \code{"five"}.}
}
\value{
If \code{x} is a factor, then \code{collapse_age()}
returns a factor; otherwise it returns a
character vector.
}
\description{
Convert age group labels to a less detailed classification.
The three classifications recognized by \code{collapse_age()}
are \code{"single"}, \code{"five"}, and \code{"lt"}, as defined on
\code{\link[=age_labels]{age_labels()}}. The following conversions are permitted:

\code{"single"} ---> \code{"lt"}
\code{"single"} ---> \code{"five"}
\code{"lt"} ---> \code{"five"}
}
\examples{
x <- c("0", "5", "3", "12")
collapse_age(x)
collapse_age(x, type_to = "lt")
}
\seealso{
\itemize{
\item \code{\link[=age_labels]{age_labels()}} to create age group labels
\item \code{\link[=clean_age]{clean_age()}} to convert existing age group labels
to a standard format
\item \code{\link[=set_age_open]{set_age_open()}} to set the lower limit
of the open age group
}
}
