% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeexp.R
\name{.lifeexp}
\alias{.lifeexp}
\title{Low-level function for calculating life expectancy
from mortality rates}
\usage{
.lifeexp(mx, age_groups, method)
}
\arguments{
\item{mx}{Mortality rates. A matrix of non-negative values.
Must have at least one column.}

\item{age_groups}{Type of age groups used. Choices are
\code{"lt"}, \code{"single"}, and \code{"five"}. See Details.}

\item{method}{Method for converting calculating
life expectancy. See Details for options.}
}
\value{
A numeric vector with length \code{ncol(mx)}.
}
\description{
Calculate life expectancy from mortality rates.
This is an internal functions, for use by
developers, and not normally be called directly
by end users.
}
\details{
Mortality rates \code{mx} are held in a matrix,
with one set of age-specific rates per column.

There are three choices for argument \code{age_groups}:
\itemize{
\item \code{"lt"}. Life table age groups \verb{"0", "1-4", "5-9", \\dots, "A+"}.
\item \code{"single"}. \verb{"0", "1", "2", \\dots, "A+"}.
\item \code{"five"}. \verb{"0-4", "5-9", \\dots, "A+"}.
}

In all three cases, the last interval is always
assumed to be open, meaning that
it includes everyone over a certain age.
}
\examples{
mx <- matrix(c(0.010, 0.002, 0.070, 0.200,
               0.011, 0.003, 0.072, 0.210),
             ncol = 2)

.lifeexp(mx, age_groups = "lt", method = "mid")
.lifeexp(mx, age_groups = "lt", method = "CD-Female")
.lifeexp(mx, age_groups = "single", method = "CD-Female")
.lifeexp(mx, age_groups = "single", method = "const")
}
\seealso{
\itemize{
\item \code{\link[=lifeexp]{lifeexp()}} is a much more user-friendly
function for calculating life expectancy.
}
}
\keyword{internal}
