% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{iran_fertility}
\alias{iran_fertility}
\title{Age-Specific Fertility Rates in Iran}
\format{
A tibble with 2010 rows and the following
columns:
\itemize{
\item time Calendar year
\item age Five-year age group from \code{"15-19"} to \code{"45-49"}
\item area \verb{"Rural" or }"Urban"`
\item rate Age-specific fertility rate
}
}
\source{
Tables 4.1 and 4.2 of
Abbasi-Shavazi, M J, McDonald, P (2005).
\emph{National and provincial level fertility trends in Iran,
1972–2006}. Australian National University.
Working Papers in Demography no. 94.
}
\usage{
iran_fertility
}
\description{
Estimates of age-specific fertility rates,
(births per 1000 person-years lived)
for rural and urban areas, in Iran, 1986-2000.
Calculated by Mohammad Jalal Abbasi-Shavazi and
Peter McDonald from data from the 2000 Iran
Demographic and Health Survey.
}
\keyword{datasets}
