% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{le}
\alias{le}
\title{Calculate life expectancy from mortality rates}
\usage{
le(mx, age_groups, method)
}
\arguments{
\item{mx}{Mortality rates. A matrix of non-negative values.
Must have at least one column.}

\item{age_groups}{Type of age groups used. Choices are
\code{"lt"}, \code{"single"}, and \code{"five"}.}

\item{method}{Method for converting calculating
life expectancy. Choices are
\code{"const"}, \code{"mid"},
\code{"CD-Female"}, \code{"CD-Male"},
\code{"HMD-Female"}, and \code{"HMD-Male"}.}
}
\value{
A numeric vector with length \code{nrow(mx)}.
}
\description{
Given mortality rates, a description of the age groups,
and a calculation method, derive life expectancies.
}
\details{
Mortality rates \code{mx} are held in a matrix,
with one set of age-specific rates per row.

There are three choices for argument \code{age_groups}:
\describe{
\item{\code{"lt"}}{Life table age groups
\code{"0", "1-4", "5-9", \dots, "A+"}}
\item{\code{"single"}}{\code{"0", "1", "2", \dots, "A+"}}
\item{\code{"five"}}{\code{"0-4", "5-9", \dots, "A+"}}
}
The last interval is always assumed to be open, meaning that
it includes everyone over a certain age.

There are six choices for the \code{method} argument.
Each method makes different assumptions about
the shape of mortality rates, the survival curve,
or the average of deaths within each age interval:
\describe{
\item{\code{"const"}}{Mortality rates are constant
within each age group; equivalently, the
life table function \code{lx} is an exponential
curve within each age interval.}
\item{\code{"mid"}}{On average, people die half way
through each age group; equivalently, the
life table function \code{lx} is a straight
line within each age group.}
\item{\code{"CD-Female"}, \code{"CD-Male"}}{As for
\code{"mid"}, except that the average age at
which infants die (and, if \code{age_groups} is
\code{"lt"}, the average age at which
children aged 1-4 die), is determined by
formulas developed by Coale and Demeny (1983),
and reported in Preston et al (2001).}
\item{\code{"HMD-Female"}, \code{"HMD-Male"}}{The
approach used in the Human Mortality Database.
As for \code{"mid"}, except that the average
age at which infants die is determined by
formulas developed by Andreev and Kingkade (2015),
and reported in Wilmoth et al (2019).}
}
Methods \code{"const"} and \code{"mid"} can be used
with any sex/gender.
Methods \code{"CD-Female"} and \code{"HMD-Female"}
should only be used for mortality rates for females,
and \code{"CD-Female"} and \code{"HMD-Male"}
should only be used for mortality rates for males.

Some methods cannot be applied to some ways of
defininge age groups. The permitted combinations are:\tabular{llll}{
    \tab \code{"lt"} \tab \code{"single"} \tab \code{"five"} \cr
   \code{"const"} \tab Yes \tab Yes \tab Yes \cr
   \code{"mid"} \tab Yes \tab Yes \tab Yes \cr
   \code{"CD-Female"} \tab Yes \tab Yes \tab  \cr
   \code{"CD-Male"} \tab Yes \tab Yes \tab  \cr
   \code{"HMD-Female"} \tab  \tab Yes \tab  \cr
   \code{"HMD-Male"} \tab  \tab Yes \tab  \cr
}
}
\examples{
mx <- matrix(c(0.010, 0.002, 0.070, 0.200,
               0.011, 0.003, 0.072, 0.210),
             nrow = 2)

le(mx, age_groups = "lt", method = "mid")
le(mx, age_groups = "lt", method = "CD-Female")
le(mx, age_groups = "single", method = "CD-Female")
le(mx, age_groups = "single", method = "const")
}
