% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetab.R
\name{lifetab}
\alias{lifetab}
\alias{lifeexp}
\title{NO_TESTS
Calculate life tables or life expectancies}
\usage{
lifetab(
  data,
  mx,
  age = NULL,
  sex = NULL,
  ax = NULL,
  by = NULL,
  method = c("const", "mid", "CD", "HMD"),
  at = 0,
  prefix = NULL
)

lifeexp(
  data,
  mx,
  age = NULL,
  sex = NULL,
  ax = NULL,
  by = NULL,
  method = c("const", "mid", "CD", "HMD"),
  at = 0,
  prefix = NULL
)
}
\arguments{
\item{data}{Data frame with mortality data.}

\item{mx}{<\code{\link[tidyselect:language]{tidyselect}}>
Mortality rates. Possibly an \link[rvec:rvec]{rvec}.}

\item{age}{<\code{\link[tidyselect:language]{tidyselect}}>
Age group labels. The labels must be
interpretable by functions
such as \code{\link[=reformat_age]{reformat_age()}} and \code{\link[=age_group_type]{age_group_type()}}.
The first age group must start at age 0, and the
last age group must be "open", with no upper
limit.}

\item{sex}{<\code{\link[tidyselect:language]{tidyselect}}>
Biological sex, with labels that can be
interprted by \code{\link[=reformat_sex]{reformat_sex()}}. Needed only when
(i) \code{infant} is \code{"CD"} or \code{"HMD"}, or
(ii) \code{age} includes age group \code{"1-4"} and
\code{child} is \code{"CD"}.}

\item{ax}{<\code{\link[tidyselect:language]{tidyselect}}>
Average age at death within age group.
Optional. See Details.}

\item{by}{<\code{\link[tidyselect:language]{tidyselect}}>
Separate life tables, or life expectancies,
calculated for each combination the \code{by} variables.
Any \code{sex} argument automatically included
in the \code{by} variables.}

\item{prefix}{Optional prefix added to new
columns in result.}

\item{infant}{Method used to calculate
life table values in age group \code{"0"}.
See Details. Default is \code{"const"}.}

\item{child}{Method used to calculate
life table values in age group \code{"1-4"}.
See Details. Default is \code{"const"}.}

\item{open}{Method used to calculate
life table values in open age group.
See Details. Default is `"const".}

\item{radix}{}
}
\value{
A \link[tibble:tibble]{tibble}.
}
\description{
Transform estimates of mortality rates
into life table quantities. Function
\code{lifetab()} returns the full life table.
Function \code{lifeexp()} life expectancy at birth.
}
\section{Calculation methods}{


There are four choices for the \code{method} argument.
Each method makes different assumptions about
the way that mortality rates vary within age group,
\itemize{
\item \code{"const"}. Mortality rates are constant
within each age group; equivalently, the
life table function \code{lx} is an exponential
curve within each age interval.
\item \code{"mid"}. On average, people die half way
through each age group; equivalently, the
life table function \code{lx} is a straight
line within each age group.
\item \code{"CD"}. As for \code{"mid"}, except that the average age at
which infants die (and, if \code{age_groups} is
\code{"lt"}, the average age at which
children aged 1-4 die), is determined by
formulas developed by Coale and Demeny (1983),
and reported in Preston et al (2001).
\item \code{"HMD"}. The  approach used in
the Human Mortality Database.
As for \code{"mid"}, except that the average
age at which infants die is determined by
formulas developed by Andreev and Kingkade (2015),
and reported in Wilmoth et al (2019).
}

Some methods cannot be applied to some
types of age groups:\tabular{llll}{
    \tab \code{"single"} \tab \code{"five"} \tab \code{"lt"} \cr
   \code{"const"} \tab Yes \tab Yes \tab Yes \cr
   \code{"mid"} \tab Yes \tab Yes \tab Yes \cr
   \code{"CD"} \tab Yes \tab No \tab Yes \cr
   \code{"HMD"} \tab Yes \tab No \tab No \cr
}
}

\section{ax}{


Average number of years lived within age group
by people who die within the age group: sometimes
referred to as a 'separation factor'. If a non-\code{NA}
value is supplied for an age group, then
calculations based on the identity

\deqn{m_x = d_x / (n_x l_x + a_x d_x)}

(where \eqn{m_x} is the mortality rate,
\eqn{d_x} is deaths, \eqn{n_x} is the width of the
age group, and \eqn{l_x} is the probability
of surviving to age \eqn{x})
are used for that age group, regardless of the
value for \code{"method"}. Can be used to fine-tune
methods (by suppling only a few non-\code{NA} values),
or to create life tables or life expectancies based
entirely on \code{ax} (by suppliying non-\code{NA} values for
all age groups.) See below for examples.
}

\references{
\itemize{
\item Preston S H, Heuveline P, and Guillot M. 2001.
\emph{Demography: Measuring and Modeling Population Processes}
Oxford: Blackwell.
\item Human Mortality Database \href{https://www.mortality.org/File/GetDocument/Public/Docs/MethodsProtocolV6.pdf}{Methods Protocol}.
\item IUSSP \href{https://demographicestimation.iussp.org}{Tools for Demographic Estimation}.
}
}
