% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{matrix_to_list_of_cols}
\alias{matrix_to_list_of_cols}
\alias{matrix_to_list_of_rows}
\title{Turn a matrix into a list of columns or rows}
\usage{
matrix_to_list_of_cols(m)

matrix_to_list_of_rows(m)
}
\arguments{
\item{m}{A matrix}
}
\value{
A list of vectors.
}
\description{
Given a matrix, create a list, each element of which contains
}
\details{
\itemize{
\item a column from the matrix (\code{matrix_to_list_of_cols()}), or
\item a row from the matrix (\code{matrix_to_list_of_rows()}).
}

\code{matrix_to_list_of_cols()} and `matrix_to_list_of_rows() are
internal functions, for use by developers, and would not
normally be called directly by end users.
}
\examples{
m <- matrix(1:12, nrow = 3)
matrix_to_list_of_cols(m)
matrix_to_list_of_rows(m)
}
