% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nzmort}
\alias{nzmort}
\title{Mortality Data for New Zealand}
\format{
A data frame with 84 rows and the
following variables:
\itemize{
\item \code{year}: Calendar year.
\item \code{gender}: \code{"Female"}, and \code{"Male"}.
\item \code{age}: Age, in life table age groups, with an open age
group of 95+.
\item \code{deaths}: Counts of deaths, randomly rounded to base 3.
\item \code{popn}: Estimates of average annual population.
\item \code{mx}: Mortality rates (deaths / popn).
}
}
\source{
Modified from data in tables
"Deaths by age and sex (Annual-Dec)" and
"Estimated Resident Population by Age and Sex (1991+) (Annual-Dec)"
from Stats NZ online database \emph{Infoshare},
downloaded on 24 September 2023.
}
\usage{
nzmort
}
\description{
Counts of deaths and population, by age, sex, and calendar
year, plus mortality rates, for New Zealand, 2021-2022.
}
\keyword{datasets}
