% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nzmort_rvec}
\alias{nzmort_rvec}
\title{Mortality Data and Probabilistic Rates for New Zealand}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 84 rows and 4 columns.
}
\usage{
nzmort_rvec
}
\description{
A modified version of \code{link{nzmort}} where \code{mx}
columns is an \link[rvec:rvec]{rvec}, rather than an ordinary
R vector. The rvec holds the random draws from the posterior
distribution obtained from by a Bayesian statistical model.
}
\keyword{datasets}
