% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poputils-package.R
\docType{package}
\name{poputils-package}
\alias{poputils}
\alias{poputils-package}
\title{Functions for manipulating and analysing
demographic data}
\description{
Miscellaneous functions for demographic analysis. Includes functions for life tables, and working with age labels.
}
\author{
\strong{Maintainer}: John Bryant \email{john@bayesiandemography.com}

}
\keyword{internal}
