% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfr.R
\name{tfr}
\alias{tfr}
\title{Calculate Total Fertility Rates}
\usage{
tfr(
  data,
  rates = NULL,
  age = age,
  sex = NULL,
  by = NULL,
  denominator = 1,
  suffix = NULL
)
}
\arguments{
\item{data}{Data frame with age-specific fertility rates
and age}

\item{rates}{Age-specific fertility rates.
Possibly an \link[rvec:rvec]{rvec}.}

\item{age}{<\code{\link[tidyselect:language]{tidyselect}}>
Age group labels. The labels must be
interpretable by functions
such as \code{\link[=reformat_age]{reformat_age()}} and \code{\link[=age_group_type]{age_group_type()}}.
The age groups must not have gaps,
and the highest age group must be "closed"
(ie have an upper limit.)}

\item{sex}{<\code{\link[tidyselect:language]{tidyselect}}>
Biological sex, with labels that can be
interpreted by \code{\link[=reformat_sex]{reformat_sex()}}.}

\item{by}{<\code{\link[tidyselect:language]{tidyselect}}>
Separate total fertility rates are calculated
for each combination the \code{by} variables.
If \code{data} is a
\href{https://dplyr.tidyverse.org/reference/group_data.html}{grouped}
data frame, then the grouping variables
take precedence over \code{by}.}

\item{denominator}{The denominator used to
calculate the rates. Default is 1.}

\item{suffix}{Optional suffix added to \code{"tfr"}
column in result.}
}
\value{
A \link[tibble:tibble]{tibble}.
}
\description{
Calculate Total Fertility Rates
}
\section{Sex or gender variable}{


Not used in official publications, but maybe in model.
}

\section{Denominator}{


1000 is common.
}

\section{Using rvecs to represent uncertainty}{


An \link[rvec:rvec]{rvec} is a 'random vector',
holding multiple draws from a distribution.
Using an rvec for the \code{mx} argument to
\code{lifetab()} or \code{lifeexp()} is a way of representing
uncertainty. This uncertainty is propagated
through to the life table values, which will
also be rvecs.
}

\seealso{
\itemize{
\item \code{\link[=lifeexp]{lifeexp()}} Calculate life expectancy from age-specific
mortality rates.
}
}
