% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeexp.R
\name{lifeexp}
\alias{lifeexp}
\title{Calculate life expectancy from a matrix of
mortality rates}
\usage{
lifeexp(mx, age, method)
}
\arguments{
\item{mx}{Mortality rates. A matrix of non-negative values.
Must have at least one column.}

\item{age}{Type of age groups used. Choices are
\code{"lt"}, \code{"single"}, \code{"five"}.}

\item{method}{Name of method for converting mortality
rates to probabilies of dying. Choices are \code{"mid"},
\code{"CD-Female"}, \code{"CD-Male"},
\code{"HMD-Female"}, \code{"HMD-Male"},
and \code{"const"}.}
}
\value{
A numeric vector with length \code{nrow(mx)}.
}
\description{
Calculate life expectancy.
}
\details{
\describe{
\item{\code{"mid"}}{ On average, people die half way
through each age interval; equivalently, the
life table function \code{lx} is a straight
line within each age interval. }
\item{\code{"CD-Female"}, \code{"CD-Male"}}{ As for
\code{"mid"}, except that the average age at
which infants die (and, if \code{age} is
\code{"lt"}, the average age at which
children aged 1-4 die), is determined by
formulas developed by Coale and Demeny, and
reported in Preston et al (2001). }
\item{\code{"HMD-Female"}, \code{"HMD-Male"}}{ The
approach used in the Human Mortality Database.
As for \code{"mid"}, except that the average
age at which infants die is determined by
formulas developed by , and reported in Preston
et al (2001). }
\item{\code{"const"}}{ Mortality rates are constant
within each age interval; equivalently, the
life table function \code{lx} is an exponential
curve within each age interval. }
}\tabular{llll}{
    \tab \code{"lt"} \tab \code{"single"} \tab \code{"five"} \cr
   \code{"mid"} \tab X \tab X \tab X \cr
   \code{"CD-Female"} \tab X \tab X \tab  \cr
   \code{"CD-Male"} \tab X \tab X \tab  \cr
   \code{"HMD-Female"} \tab  \tab X \tab  \cr
   \code{"HMD-Male"} \tab  \tab X \tab  \cr
   \code{"const"} \tab X \tab X \tab X \cr
}


The last age group is always the open age group.

If \code{age} is \code{"lt"}, then \code{mx} must have
at least three columns (ie at least three age groups).
}
\examples{
mx <- matrix(c(0.010, 0.002, 0.070, 0.200,
               0.011, 0.003, 0.072, 0.210),
             nrow = 2)

lifeexp(mx, age = "lt", method = "mid")
lifeexp(mx, age = "lt", method = "CD-Female")
lifeexp(mx, age = "single", method = "CD-Female")
lifeexp(mx, age = "single", method = "const")
}
