% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetab.R
\name{lifetab}
\alias{lifetab}
\alias{lifeexp}
\title{HAS_TESTS
Calculate life tables or life expectancies}
\usage{
lifetab(
  data,
  mx = mx,
  age = age,
  sex = NULL,
  ax = NULL,
  by = NULL,
  infant = c("constant", "linear", "CD", "HMD"),
  child = c("constant", "linear", "CD"),
  closed = c("constant", "linear"),
  open = "constant",
  radix = 1e+05,
  prefix = NULL
)

lifeexp(
  data,
  mx = mx,
  age = age,
  sex = NULL,
  ax = NULL,
  by = NULL,
  infant = c("constant", "linear", "CD", "HMD"),
  child = c("constant", "linear", "CD"),
  closed = c("constant", "linear"),
  open = "constant",
  prefix = NULL
)
}
\arguments{
\item{data}{Data frame with mortality data.}

\item{mx}{<\code{\link[tidyselect:language]{tidyselect}}>
Mortality rates. Possibly an \link[rvec:rvec]{rvec}.}

\item{age}{<\code{\link[tidyselect:language]{tidyselect}}>
Age group labels. The labels must be
interpretable by functions
such as \code{\link[=reformat_age]{reformat_age()}} and \code{\link[=age_group_type]{age_group_type()}}.
The first age group must start at age 0, and the
last age group must be "open", with no upper
limit.}

\item{sex}{<\code{\link[tidyselect:language]{tidyselect}}>
Biological sex, with labels that can be
interprted by \code{\link[=reformat_sex]{reformat_sex()}}. Needed only when
\code{infant} is \code{"CD"} or \code{"HMD"}, or \code{child} is
\code{"CD"}.}

\item{ax}{<\code{\link[tidyselect:language]{tidyselect}}>
Average age at death within age group.
Optional. See Details.}

\item{by}{<\code{\link[tidyselect:language]{tidyselect}}>
Separate life tables, or life expectancies,
calculated for each combination the \code{by} variables.
Any \code{sex} argument automatically included
in the \code{by} variables.}

\item{infant}{Method used to calculate
life table values in age group \code{"0"}.
Ignored if \code{age} does not include age group \code{"0"}.
Default is \code{"constant"}.}

\item{child}{Method used to calculate
life table values in age group \code{"1-4"}.
Ignored if \code{age} does not include age group \code{"0"}.
Default is \code{"constant"}.}

\item{closed}{Method used to calculate
life table values in closed age intervals
other than \code{"0"} and \code{"1-4"} (ie intervals
such as "10-14" or "12"). Default is \code{"constant"}.}

\item{open}{Method used to calculate
life table values in the final, open age group
(eg \code{"80+"} or \code{"110+"}).
Currently the only option is `"constant".}

\item{radix}{Initial population for the
\code{lx} column. Default is \code{100000}.}

\item{prefix}{Optional prefix added to new
columns in result.}
}
\value{
A \link[tibble:tibble]{tibble}.
}
\description{
Transform estimates of mortality rates
into life table quantities. Function
\code{lifetab()} returns the full life table.
Function \code{lifeexp()} life expectancy at birth.
}
\section{Definitions of life table quantities}{

\itemize{
\item \code{qx} Probability of surviving from the start
of age group 'x' to the end.
\item \code{lx} Number of people alive at
the start of age group \code{x}.
\item \code{dx} Number of deaths in age group \code{x}
\item \code{Lx} Expected number of person years lived in
age group \code{x}.
\item \code{ex} Life expectancy, calculated at the start
of age group \code{x}.
}
}

\section{Calculation methods}{


\code{lifetab()} and \code{lifeexp()} implement a number
of different methods for
calculating life table quantities
from mortality rates. Each method makes
different assumptions about
the way that mortality rates vary within
age intervals:
\itemize{
\item \code{"constant"} Mortality rates are constant
within each interval.
\item \code{"linear"}. Life table quantity \code{lx}
is a straight line within each interval:
equivalently, deaths are distributed uniformly
within each interval.
\item \code{"CD"}. Used only with age groups "0"
and "1-4". Mortality rates decline
over the  age interval,
with the slope depending on the absolute
level of infant mortality. The formulas were
developed by Coale and Demeny (1983),
and used in Preston et al (2001).
\item \code{"HMD"}. Used only with age group "0"
(which is specified by the \code{infant} argument.)
Mortality rates decline over the age interval,
with the slope depending on the absolute
level of infant mortality. The formulas
were formulas developed by Andreev and Kingkade (2015),
and are used in the Human Mortality Database
\href{https://www.mortality.org/File/GetDocument/Public/Docs/MethodsProtocolV6.pdf}{methods protocol}.
}

For a detailed description of the methods,
see the vignette GIVE LINK.
}

\section{ax}{


\code{ax} is the average number of years
lived in an age interval by people who
die in that interval. Demographers sometimes
refer to \code{ax} as a 'separation factor'. If a non-\code{NA}
value of \code{ax} is supplied for an age group,
then the results for that age group are based
on the formula

\deqn{m_x = d_x / (n_x l_x + a_x d_x)},

(where \code{n_x} is the width of the age interval)
over-riding any methods specified via the \code{infant}, \code{child},
\code{closed} and \code{open} arguments.
}

\section{Using rvecs to represent uncertainty}{


An \link[rvec:rvec]{rvec} is a 'random vector',
holding multiple draws from a distribution.
Using an rvec for the \code{mx} argument to
\code{lifetab()} or \code{lifeexp()} is a way of representing
uncertainty. This uncertainty is propogated
through to the life table values, which will
also be rvecs.
}

\examples{
library(dplyr)

## life table for females based on 'level 1'
## mortality rates "West" model life table
west_mx \%>\%
    filter(sex == "Female",
           level == 1) \%>\%
    lifetab()

## change method for infant and children from
## default ("constant") to "CD"
west_mx \%>\%
    filter(sex == "Female",
           level == 1) \%>\%
    lifetab(sex = sex,
            infant = "CD",
            child = "CD")

## calculate life expectancies
## for all levels, using the 'by'
## argument to distinguish levels
west_mx \%>\%
    lifeexp(sex = sex,
            infant = "CD",
            child = "CD",
            by = level)

## obtain the same result using
## 'group_by'
west_mx \%>\%
  group_by(level) \%>\%
  lifeexp(sex = sex,
          infant = "CD",
          child = "CD")
}
\references{
\itemize{
\item Preston SH, Heuveline P, and Guillot M. 2001.
\emph{Demography: Measuring and Modeling Population Processes}
Oxford: Blackwell.
\item Coale AJ, Demeny P,  and Vaughn B. 1983.
\strong{Regional model life tables and stable populations}
New York: Academic Press.
\item Andreev, E.M. and Kingkade, W.W., 2015.
Average age at death in infancy and infant mortality level:
Reconsidering the Coale-Demeny formulas at current
levels of low mortality. \strong{Demographic Research},
33, pp.363-390.
\item Human Mortality Database \href{https://www.mortality.org/File/GetDocument/Public/Docs/MethodsProtocolV6.pdf}{Methods Protocol}.
\item \href{https://demographicestimation.iussp.org}{Tools for Demographic Estimation}.
}
}
