% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetab.R
\name{lifetab}
\alias{lifetab}
\alias{lifeexp}
\title{NO_TESTS
Calculate life tables or life expectancies}
\usage{
lifetab(
  data,
  mx,
  age = NULL,
  sex = NULL,
  ax = NULL,
  by = NULL,
  method = c("const", "mid", "CD", "HMD"),
  at = 0,
  prefix = NULL
)

lifeexp(
  data,
  mx,
  age = NULL,
  sex = NULL,
  ax = NULL,
  by = NULL,
  method = c("const", "mid", "CD", "HMD"),
  at = 0,
  prefix = NULL
)
}
\arguments{
\item{data}{Data frame with mortality data.}

\item{mx}{<\code{\link[tidyselect:language]{tidyselect}}>
Mortality rates. Possibly an \link[rvec:rvec]{rvec}.}

\item{age}{<\code{\link[tidyselect:language]{tidyselect}}>
Age group labels.}

\item{sex}{<\code{\link[tidyselect:language]{tidyselect}}>
Biological sex. Needed only if method is
\code{"CD"} or \code{"HMD"}.}

\item{ax}{<\code{\link[tidyselect:language]{tidyselect}}>
Average age at death within age group.
Optional. See Details.}

\item{by}{<\code{\link[tidyselect:language]{tidyselect}}>
Separate life expectancies, or life tables, are
calculated for each combination the \code{by} variables.}

\item{method}{Method used to calculate
rates. See Details.}

\item{at}{Age at which life tables start (\code{lifetab()}),
or at which life expectancy is calculate (\code{lifeexp()}).
Default is \code{0}.}

\item{prefix}{Optional prefix added to new
columns in result.}
}
\value{
A \link[tibble:tibble]{tibble}.
}
\description{
Calculate full life tables, or just life expectancies.
}
\section{age}{


The \code{age} variable must contain labels that functions
such as \code{\link[=reformat_age]{reformat_age()}} or \code{\link[=age_groups]{age_groups()}}
can interpret, ie that can be interpreted
as single, five-year or life-table age groups. The final
age group must be "open", with no upper limit. There also
must not be any gaps between the lowest and highest
age groups.
}

\section{sex}{


A vector containing, for each set of ages, a single
string, which can be interpreted by \code{\link[=reformat_sex]{reformat_sex()}},
as either \code{"Female"} or \code{"Male"}.

The \code{sex} argument is only need in two cases:
\itemize{
\item \code{method} is \code{"CD"} and the youngest
age group is less then 5 years old.
\item \verb{method is }"HMD"\verb{and the youngest age group is}"0"`.
}

If Coale-Demeny (CD) and Human Mortality Database
formulas are produced for sexes or genders other than \code{"Female"},
and \code{"Male"}, we will add these formulas to the package.
In the meantime, \code{"const"} and \code{"mid"}
methods can be used for any sex or gender.
}

\section{ax}{


The average age at which



Average number of years lived within age group
by people who die within the age group: sometimes
referred to as a 'separation factor'. If a non-\code{NA}
value is supplied for an age group, then
calculations based on the identity

\deqn{m_x = d_x / (n_x l_x + a_x d_x)}

(where \eqn{m_x} is the mortality rate,
\eqn{d_x} is deaths, \eqn{n_x} is the width of the
age group, and \eqn{l_x} is the probability
of surviving to age \eqn{x})
are used for that age group, regardless of the
value for \code{"method"}. Can be used to fine-tune
methods (by suppling only a few non-\code{NA} values),
or to create life tables or life expectancies based
entirely on \code{ax} (by suppliying non-\code{NA} values for
all age groups.) See below for examples.
}

\section{by}{


Life tables or life expectancies are calculated
separately for each combination of the \code{by} variables.
If a \code{sex} argument is specified, then that variable
is automatically included in the \code{by} variables.
}

\section{method}{


There are four choices for the \code{method} argument.
Each method makes different assumptions about
the way that mortality rates vary within age group,
\itemize{
\item \code{"const"}. Mortality rates are constant
within each age group; equivalently, the
life table function \code{lx} is an exponential
curve within each age interval.
\item \code{"mid"}. On average, people die half way
through each age group; equivalently, the
life table function \code{lx} is a straight
line within each age group.
\item \code{"CD"}. As for \code{"mid"}, except that the average age at
which infants die (and, if \code{age_groups} is
\code{"lt"}, the average age at which
children aged 1-4 die), is determined by
formulas developed by Coale and Demeny (1983),
and reported in Preston et al (2001).
\item \code{"HMD"}. The  approach used in
the Human Mortality Database.
As for \code{"mid"}, except that the average
age at which infants die is determined by
formulas developed by Andreev and Kingkade (2015),
and reported in Wilmoth et al (2019).
}

Some methods cannot be applied to some
types of age groups:\tabular{llll}{
    \tab \code{"single"} \tab \code{"five"} \tab \code{"lt"} \cr
   \code{"const"} \tab Yes \tab Yes \tab Yes \cr
   \code{"mid"} \tab Yes \tab Yes \tab Yes \cr
   \code{"CD"} \tab Yes \tab No \tab Yes \cr
   \code{"HMD"} \tab Yes \tab No \tab No \cr
}
}

\references{
\itemize{
\item Preston S H, Heuveline P, and Guillot M. 2001.
\emph{Demography: Measuring and Modeling Population Processes}
Oxford: Blackwell.
\item Human Mortality Database \href{https://www.mortality.org/File/GetDocument/Public/Docs/MethodsProtocolV6.pdf}{Methods Protocol}.
\item IUSSP \href{https://demographicestimation.iussp.org}{Tools for Demographic Estimation}.
}
}
